test: polish acceptance harness
